refactor : server component에서 사용되지않는 chat, tether, dollar  env정보 변경에 따른 server component data fetching -> client component data fetching으로 변경
